adding authorize permission in controller

+++ app/Http/Controllers/RoleController.php
@@ -23,6 +23,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        $this->authorize('view-roles', Role::class);
         return view('pages.roles.index', [
             'roles' => Role::with('permissions')->orderByDesc('id')->paginate(5)
         ]);
@@ -35,6 +36,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        $this->authorize('create-roles', Role::class);
         return view('pages.roles.create', ListPermission::run());
     }
 
@@ -46,6 +48,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create-roles', Role::class);
         $this->roleService->store($request);
         return redirect()->route('roles.index')->with('success', 'Role created successfully');
     }
@@ -69,6 +72,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('edit-roles', Role::class);
         $role = ['role' => new RoleResource($role)];
         return view('pages.roles.edit', array_merge($role, ListPermission::run()));
     }
@@ -94,6 +98,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        $this->authorize('delete-roles', Role::class);
         $this->roleService->delete($role);
         return redirect()->route('roles.index')->with('success', 'Role has been deleted.');
     }
